:tractor: re-organize repo related functionality
