+++ moban/mobanfile/__init__.py
@@ -8,7 +8,8 @@ from lml.utils import do_import
 import moban.reporter as reporter
 import moban.constants as constants
 from moban import plugins
-from moban.utils import merge, git_clone, pip_install
+from moban.repo import git_clone
+from moban.utils import merge, pip_install
 from moban.mobanfile.targets import parse_targets
 from moban.deprecated import deprecated
 from moban.definitions import GitRequire
