+++ tests/test_engine.py
@@ -26,7 +26,7 @@ def test_expand_pypi_dir():
         assert os.path.exists(directory)
 
 
-@patch("moban.utils.get_moban_home", return_value=USER_HOME)
+@patch("moban.repo.get_moban_home", return_value=USER_HOME)
 @patch("os.path.exists", return_value=True)
 def test_expand_repo_dir(_, __):
     dirs = list(expand_template_directories("git_repo:template"))
